New 'make release' option for modules
git-svn-id: https://barricada.hi.inet/repositorio/samson/trunk@1536 4143565c-f3ec-42ea-b729-f8ce0cf5cbc3

+++ modules/cdr/cdr_parse_fields.h
@@ -30,7 +30,7 @@
 		///
 		///@param strCellId Pointer to the string cell id.
 		///@param cellId Pointer to the integer cell id.
-		void cdrStrCellToInt( char *strCellId, unsigned int *cellId );
+		void cdrStrCellToInt( char *strCellId, size_t *cellId );
 
 		inline void GST_get_cdrs_number( char *line, size_t *num,  int pos_field )
 		{
