+++ modules/mob/operations/map_get_month.h
@@ -47,7 +47,7 @@ public:
 
 		samson::system::Void void_data;
 
-		for (int i = 0 ; i < inputs[0].num_kvs ; ++i)
+		for (size_t i = 0 ; i < inputs[0].num_kvs ; ++i)
 		{
 			//Parsing key and value
 			date.parse( inputs[0].kvs[i]->key );
