+++ modules/mob/operations/map_place_cell.h
@@ -44,7 +44,7 @@ public:
 		Place place;
 		samson::system::UInt cellId;
 
-		for (int i = 0 ; i < inputs[0].num_kvs ; ++i)
+		for (size_t i = 0 ; i < inputs[0].num_kvs ; ++i)
 		{
 			//Parsing key and value
 			place.parse( inputs[0].kvs[i]->value );
