+++ modules/sna/operations/gen_comm_clique.h
@@ -29,7 +29,7 @@ namespace sna{
 			  //output
 			  Clique clique;
 
-			  for (int i = 0; i < inputs[0].num_kvs; i++) {
+			  for (size_t i = 0; i < inputs[0].num_kvs; i++) {
 			    id.parse(inputs[0].kvs[i]->key );
 			    data.parse(inputs[0].kvs[i]->value );
