+++ modules/sna/operations/select_key_in_cliques_graph.h
@@ -28,7 +28,7 @@ namespace sna{
 			Clique clique;
 			samson::system::Void void_data;
 
-			for (int i = 0 ; i < inputs[0].num_kvs ;i++)
+			for (size_t i = 0 ; i < inputs[0].num_kvs ;i++)
 			{
 				clique.parse(inputs[0].kvs[i]->key);
 				writer->emit(0, &clique, &void_data);
