+++ modules/sna/operations/spread_cliques_per_node.h
@@ -28,7 +28,7 @@ namespace sna{
 			if ( inputs[0].num_kvs == 0)
 				return;
 
-			for (int i = 0  ; i < inputs[0].num_kvs ; i++)
+			for (size_t i = 0  ; i < inputs[0].num_kvs ; i++)
 			{
 				clique.parse( inputs[0].kvs[i]->key );
