+++ modules/sort/operations/reduce_identity.h
@@ -25,7 +25,7 @@ namespace sort{
 		void run(  samson::KVSetStruct* inputs , samson::KVWriter *writer )
 		{
 			//OLM_T(LMT_User01, ("reduce_identity: inputs[0].num_kvs:%lu", inputs[0].num_kvs));
-			for (int i = 0; (i < inputs[0].num_kvs); ++i)
+			for (size_t i = 0; (i < inputs[0].num_kvs); ++i)
 			{
 				val.parse(inputs[0].kvs[i]->key);
 				writer->emit(0, &val, &dummy);
