+++ modules/url/operations/exportHitsPerCategoryWithStrings.h
@@ -43,7 +43,7 @@ namespace url{
 			char *p_output;
 			int escritos;
 
-			for (int i = 0; (i < inputs[0].num_kvs); i++)
+			for (size_t i = 0; (i < inputs[0].num_kvs); i++)
 			{
 
 				catId.parse(inputs[0].kvs[i]->key);
