+++ modules/url/operations/exportHitsRecommendationPerUser.h
@@ -44,7 +44,7 @@ parserOut exportHitsRecommendationPerUser
 
 			int escritos;
 
-			for (int i = 0; (i < inputs[0].num_kvs); i++)
+			for (size_t i = 0; (i < inputs[0].num_kvs); i++)
 			{
 				userId.parse(inputs[0].kvs[i]->key);
 				paths.parse(inputs[0].kvs[i]->value);
